independent_core/test_with_fallbacks.py
#!/usr/bin/env python3
"""
Test script demonstrating how to use accuracy tracking modules with fallbacks
"""

import sys
import os

# Add current directory to path
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

print("="*60)
print("ACCURACY TRACKING SYSTEM - FALLBACK TEST")
print("="*60)

# Test 1: Import and check dependencies
print("\n1. Testing dependency checker...")
try:
    from dependency_checker import (
        dependency_manager, 
        print_dependency_report,
        jwt,
        toml
    )
    print("✓ Dependency checker imported successfully")
    
    # Print dependency report
    print_dependency_report()
    
except Exception as e:
    print(f"✗ Failed to import dependency checker: {e}")
    sys.exit(1)

# Test 2: Test JWT functionality
print("\n2. Testing JWT functionality...")
try:
    # Test JWT encoding/decoding
    test_payload = {
        'user_id': 'test_user',
        'role': 'admin',
        'exp': 1234567890
    }
    
    token = jwt.encode(test_payload, 'secret_key', algorithm='HS256')
    print(f"✓ JWT encode successful: {token[:20]}...")
    
    decoded = jwt.decode(token, 'secret_key', algorithms=['HS256'])
    print(f"✓ JWT decode successful: user_id={decoded.get('user_id')}")
    
except Exception as e:
    print(f"⚠️  JWT functionality limited: {e}")
    print("  Using fallback implementation")

# Test 3: Test TOML functionality
print("\n3. Testing TOML functionality...")
try:
    # Test TOML parsing
    toml_content = """
    [database]
    host = "localhost"
    port = 5432
    
    [cache]
    enabled = true
    size = 1000
    """
    
    config = toml.loads(toml_content)
    print(f"✓ TOML parsing successful: {list(config.keys())}")
    
    # Test TOML dumping
    toml_output = toml.dumps(config)
    print("✓ TOML dumping successful")
    
except Exception as e:
    print(f"⚠️  TOML functionality limited: {e}")
    print("  Using fallback implementation")

# Test 4: Test config loader
print("\n4. Testing config loader...")
try:
    # Test if we can import accuracy tracking diagnostics
    import accuracy_tracking_diagnostics
    print("✓ accuracy_tracking_diagnostics imported successfully")
    
    # Check if JWT and TOML are available in the module
    if hasattr(accuracy_tracking_diagnostics, 'jwt') and accuracy_tracking_diagnostics.jwt:
        print("✓ JWT available in accuracy_tracking_diagnostics")
    else:
        print("⚠️  JWT using fallback in accuracy_tracking_diagnostics")
    
    if hasattr(accuracy_tracking_diagnostics, 'toml') and accuracy_tracking_diagnostics.toml:
        print("✓ TOML available in accuracy_tracking_diagnostics")
    else:
        print("⚠️  TOML using fallback in accuracy_tracking_diagnostics")
    
    # Test dependency status reporting
    if hasattr(accuracy_tracking_diagnostics, 'get_dependency_status'):
        status = accuracy_tracking_diagnostics.get_dependency_status()
        print(f"✓ Dependency status: {status}")
    
except Exception as e:
    print(f"✗ Config loader test failed: {e}")

# Test 5: Test Brain system integration
print("\n5. Testing Brain system integration...")
try:
    # Test if we can import Brain system components
    from brain import Brain
    print("✓ Brain system imported successfully")
    
    # Test if dependency checker works with Brain
    brain = Brain()
    print("✓ Brain instance created successfully")
    
    # Test if we can access dependency status from Brain context
    from dependency_checker import get_dependency_status
    dep_status = get_dependency_status()
    print(f"✓ Dependency status accessible: {len(dep_status)} dependencies checked")
    
except Exception as e:
    print(f"⚠️  Brain system integration limited: {e}")
    print("  Core Brain functionality should still work")

# Test 6: Demonstrate fallback usage
print("\n6. Demonstrating fallback usage...")

# Example: Using config with different formats
print("\nConfig format support:")
formats = {
    'JSON': True,  # Always available (built-in)
    'YAML': 'yaml' in sys.modules,
    'TOML': hasattr(toml, 'load') and not hasattr(toml, '_parse_value')  # Check if real toml
}

for fmt, available in formats.items():
    status = "✓ Native" if available else "⚠️  Fallback"
    print(f"  {fmt}: {status}")

# Example: Creating configs with available formats
print("\nCreating test configurations:")

# JSON config (always works)
import json
json_config = {"test": "json_value"}
json_str = json.dumps(json_config)
print(f"  ✓ JSON config: {json_str}")

# YAML config (with fallback)
try:
    yaml_config = {"test": "yaml_value"}
    if 'yaml' in sys.modules:
        import yaml
        yaml_str = yaml.dump(yaml_config)
        print(f"  ✓ YAML config: {yaml_str.strip()}")
    else:
        print("  ⚠️  YAML using fallback")
except:
    print("  ⚠️  YAML using fallback")

# TOML config (with fallback)
try:
    toml_config = {"test": "toml_value"}
    toml_str = toml.dumps(toml_config)
    print(f"  ✓ TOML config: {toml_str.strip()}")
except:
    print("  ⚠️  TOML using fallback")

# Summary
print("\n" + "="*60)
print("SUMMARY")
print("="*60)
print("\n✓ The accuracy tracking system can run with fallbacks")
print("✓ Missing dependencies will use limited implementations")
print("✓ For full functionality, install all dependencies:")
print("  pip install -r requirements.txt")
print("\n⚠️  Fallback implementations have limitations:")
print("  - JWT: Basic encoding/decoding only")
print("  - TOML: Simple key-value parsing")
print("  - YAML: Basic structure parsing")
print("  - Encryption: Insecure XOR (not for production)")
print("="*60)

# Test if we can create a minimal working setup
print("\n7. Creating minimal working setup...")
try:
    # This demonstrates that the system can work even with fallbacks
    from pathlib import Path
    
    # Create test config file
    test_config_path = Path("test_config.json")
    with open(test_config_path, 'w') as f:
        json.dump({
            "version": "1.0.0",
            "environment": "development",
            "api": {"host": "localhost", "port": 8000}
        }, f)
    
    # Load config
    with open(test_config_path, 'r') as f:
        loaded_config = json.load(f)
    
    print(f"✓ Created and loaded test config: {loaded_config['version']}")
    
    # Clean up
    test_config_path.unlink()
    
    print("✓ Minimal setup works with available dependencies")
    
except Exception as e:
    print(f"✗ Minimal setup failed: {e}")

# Test 8: Demonstrate JWT fallback functionality
print("\n8. Testing JWT fallback robustness...")
try:
    # Test various JWT scenarios
    test_cases = [
        {'user': 'admin', 'role': 'administrator'},
        {'session_id': '12345', 'timestamp': 1640995200},
        {'permissions': ['read', 'write'], 'active': True}
    ]
    
    for i, payload in enumerate(test_cases, 1):
        try:
            token = jwt.encode(payload, f'secret_{i}', algorithm='HS256')
            decoded = jwt.decode(token, f'secret_{i}', algorithms=['HS256'])
            print(f"  ✓ Test case {i}: {list(payload.keys())} -> {list(decoded.keys())}")
        except Exception as e:
            print(f"  ✗ Test case {i} failed: {e}")
    
except Exception as e:
    print(f"✗ JWT fallback testing failed: {e}")

# Test 9: Demonstrate TOML fallback functionality  
print("\n9. Testing TOML fallback robustness...")
try:
    # Test various TOML structures
    test_configs = [
        '[server]\nhost = "localhost"\nport = 8000',
        '[database]\nname = "test"\nuser = "admin"\npassword = "secret"',
        '[features]\nenabled = true\ndebug = false\nmax_connections = 100'
    ]
    
    for i, toml_str in enumerate(test_configs, 1):
        try:
            parsed = toml.loads(toml_str)
            reconstructed = toml.dumps(parsed)
            print(f"  ✓ TOML test {i}: {len(parsed)} sections parsed and reconstructed")
        except Exception as e:
            print(f"  ✗ TOML test {i} failed: {e}")
    
except Exception as e:
    print(f"✗ TOML fallback testing failed: {e}")

print("\n✅ Test completed. System is functional with fallbacks.\n")

# Final integration test
print("10. Final integration test...")
try:
    # Test that everything works together
    from dependency_checker import print_dependency_report
    
    print("\nFinal dependency report:")
    print_dependency_report()
    
    print("\n✅ All integration tests completed successfully!")
    print("The system is ready to use with available dependencies.")
    
except Exception as e:
    print(f"✗ Final integration test failed: {e}")
    print("Some components may not work correctly.")

print("\n" + "="*60)